# Proposed optimizations to BAM records