--- a/package.json
+++ b/package.json
@@ -39,34 +39,29 @@
   ],
   "dependencies": {
     "@gmod/bgzf-filehandle": "^1.3.3",
-    "abortable-promise-cache": "^1.4.0",
     "buffer-crc32": "^0.2.13",
     "cross-fetch": "^3.0.2",
-    "es6-promisify": "^6.0.1",
     "generic-filehandle": "^2.0.0",
     "long": "^4.0.0",
-    "object.entries-ponyfill": "^1.0.1",
-    "quick-lru": "^2.0.0",
-    "range-parser": "^1.2.1"
+    "object.entries-ponyfill": "^1.0.1"
   },
   "devDependencies": {
     "@types/buffer-crc32": "^0.2.0",
-    "@types/jest": "^27.0.2",
+    "@types/jest": "^27.4.1",
     "@types/long": "^4.0.0",
-    "@types/node": "^12.7.8",
-    "@types/range-parser": "^1.2.3",
-    "@typescript-eslint/eslint-plugin": "^5.7.0",
-    "@typescript-eslint/parser": "^5.7.0",
-    "eslint": "^7.0.0",
-    "eslint-config-prettier": "^8.3.0",
-    "eslint-plugin-import": "^2.25.3",
+    "@types/node": "^17.0.21",
+    "@typescript-eslint/eslint-plugin": "^5.12.1",
+    "@typescript-eslint/parser": "^5.12.1",
+    "eslint": "^8.9.0",
+    "eslint-config-prettier": "^8.4.0",
+    "eslint-plugin-import": "^2.25.4",
     "eslint-plugin-prettier": "^4.0.0",
-    "jest": "^27.3.1",
+    "jest": "^27.5.1",
     "prettier": "^2.4.1",
     "rimraf": "^3.0.2",
-    "standard-changelog": "^1.0.0",
-    "ts-jest": "^27.0.7",
-    "typescript": "^4.4.4"
+    "standard-changelog": "^2.0.27",
+    "ts-jest": "^27.1.3",
+    "typescript": "^4.5.5"
   },
   "publishConfig": {
     "access": "public"