--- a/src/bai.ts
+++ b/src/bai.ts
@@ -62,11 +62,9 @@ export default class BAI extends IndexFile {
     const binLimit = ((1 << ((depth + 1) * 3)) - 1) / 7
 
     // read the indexes for each reference sequence
-    data.indices = new Array(data.refCount)
+    data.indices = []
     let currOffset = 8
     for (let i = 0; i < data.refCount; i += 1) {
-      await abortBreakPoint(opts.signal)
-
       // the binning index
       const binCount = bytes.readInt32LE(currOffset)
       let stats
@@ -85,7 +83,7 @@ export default class BAI extends IndexFile {
         } else {
           const chunkCount = bytes.readInt32LE(currOffset)
           currOffset += 4
-          const chunks = new Array(chunkCount)
+          const chunks = []
           for (let k = 0; k < chunkCount; k += 1) {
             const u = fromBytes(bytes, currOffset)
             const v = fromBytes(bytes, currOffset + 8)
@@ -102,7 +100,7 @@ export default class BAI extends IndexFile {
       // as we're going through the linear index, figure out
       // the smallest virtual offset in the indexes, which
       // tells us where the BAM header ends
-      const linearIndex = new Array(linearCount)
+      const linearIndex = []
       for (let k = 0; k < linearCount; k += 1) {
         linearIndex[k] = fromBytes(bytes, currOffset)
         currOffset += 8