--- a/src/declare.d.ts
+++ b/src/declare.d.ts
@@ -1,3 +1,2 @@
 declare module 'object.entries-ponyfill'
-declare module 'quick-lru'
 declare module '@gmod/bgzf-filehandle'