--- a/src/record.ts
+++ b/src/record.ts
@@ -520,21 +520,21 @@ export default class BamRecord {
       start + 36 + this.get('_l_read_name') + this.get('_n_cigar_op') * 4
     const seqBytes = this.get('_seq_bytes')
     const len = this.get('seq_length')
-    let buf = ''
+    let buf = []
     for (let j = 0; j < seqBytes - 1; ++j) {
       const sb = byteArray[p + j]
-      buf += SEQRET_DECODER[(sb & 0xf0) >> 4]
-      buf += SEQRET_DECODER[sb & 0x0f]
+      buf.push(SEQRET_DECODER[(sb & 0xf0) >> 4])
+      buf.push(SEQRET_DECODER[sb & 0x0f])
     }
 
     // there are two bases per byte, so this handles the case where it is odd
     // length
     const sb = byteArray[p + seqBytes - 1]
-    buf += SEQRET_DECODER[(sb & 0xf0) >> 4]
+    buf.push(SEQRET_DECODER[(sb & 0xf0) >> 4])
     if (len % 2 === 0) {
-      buf += SEQRET_DECODER[sb & 0x0f]
+      buf.push(SEQRET_DECODER[sb & 0x0f])
     }
-    return buf
+    return buf.join('')
   }
 
   // adapted from igv.js