The profiler says that these changes make the "serializeResultsInWorkers" about 8x faster for BAM files (800ms to 100ms). The entire chunk of steps goes from 3.39 seconds to 0.71s (~4.7x faster)

Mainly appears to be due to reduced GC pressure

This caches the quality scores on the feature, which may increase memory usage

![Screenshot from 2022-02-25 10-36-51](https://user-images.githubusercontent.com/6511937/155761394-19391c46-3c31-4926-8315-b90dd27a9890.png)
![Screenshot from 2022-02-25 10-37-03](https://user-images.githubusercontent.com/6511937/155761397-4ecf615d-c60d-479e-91ea-d44accbe95e9.png)