@rbuels 2022-02-25
wow nice work. memoizing the qualRaw seems to have helped a lot!