@cmdcolin 2022-02-25
modified qualRaw to return a slice of the underlying buffer instead of copying to a new buffer or array, which should be fast without increasing memory usage

loading this 80x coverage hifi reads in jbrowse, total amount of time to load the snpcov and pileup with just this track, time goes from 14.5 seconds to about 8 seconds with this PR 

might have overstated the 8x improvement in initial post, but it's about 2x faster for things like the serialize to json