@cmdcolin 2022-02-25
to be clear, no more memoizing